refactor: rename and relocate student modal components
